persoonal_X_AI_agent/requirements.txt
tweepy
python-dotenv
schedule 